move mock classes, move exceptions
