+++ setup.py
@@ -87,6 +87,6 @@ setup(
             "result = orquesta.expressions.functions.workflow:result_",
         ],
         "orquesta.tests": ["fake = orquesta.tests.unit.utils.test_plugin:FakePlugin"],
-        "console_scripts": ["orquesta-test = orquesta.fixture_mock:main"],
+        "console_scripts": ["orquesta-test = orquesta.test.mocks:main"],
     },
 )
